Added dummy powered rails
