+++ pumpkin/src/block/blocks/mod.rs
@@ -16,12 +16,12 @@ pub(crate) mod fences;
 pub(crate) mod furnace;
 pub(crate) mod jukebox;
 pub(crate) mod logs;
-pub(crate) mod rail;
 pub(crate) mod redstone;
 pub(crate) mod signs;
 pub(crate) mod sugar_cane;
 pub(crate) mod tnt;
 pub(crate) mod torches;
+
 /// The standard destroy with container removes the player forcibly from the container,
 /// drops items to the floor, and back to the player's inventory if the item stack is in movement.
 pub async fn standard_on_broken_with_container(
